Widen the UI vertex index, and the wrap it was refusing
`UiGeometry.Indices` was `ushort`, and `UiGeometryBuilder.Quad` refused to
emit anything past 65 535 vertices rather than wrap. Refusing was the honest
thing to do while the index was narrow; it is not a fix, because a dense
editor window really can pass sixteen thousand quads and the symptom of
dropping the rest is a frame that is missing its bottom half.

Thirty-two bits, and not because a frame is expected to need them. Almost
none do. The reason is what happens to the one that does: an index that
wraps draws geometry from the top of the frame in the middle of it, and it
does that silently. Two bytes per index on about an eighth of the frame's
bytes buys never having to reason about the ceiling again.

The test is the whole of what this bought — 20 000 quads, every one present,
and the last quad's indices naming the last four vertices. Under the old
`ushort` the count assertion catches the refusal; under a `ushort` that did
not refuse, the last-quad assertion catches the wrap. Sabotaging `start` back
to a 16-bit truncation fails it.

+++ Core/Vixen.Ui/Rendering/UiGeometry.cs
@@ -47,8 +47,17 @@ public readonly record struct UiDraw(BatchKind Kind, int First, int Count, int F
 /// <param name="Vertices">Every vertex, in painting order.</param>
 /// <param name="Indices">Triangles into <paramref name="Vertices" />.</param>
 /// <param name="Draws">What to draw, in painting order.</param>
+/// <remarks>
+///     ⚠ <b>Thirty-two-bit indices, and not because a frame is expected to need them.</b> Almost none
+///     do: sixteen bits reach 65 535 vertices, which is sixteen thousand quads, and a dense editor
+///     window is a few thousand. The reason is what happens to the one frame that does — an index
+///     that wraps draws geometry from the top of the frame in the middle of it, silently, and the
+///     picture is wrong rather than absent. Emitting the wider index costs two bytes per index on
+///     about an eighth of the frame's bytes, and buys never having to reason about the ceiling
+///     again.
+/// </remarks>
 public readonly record struct UiGeometry(
     IReadOnlyList<UiVertex> Vertices,
-    IReadOnlyList<ushort> Indices,
+    IReadOnlyList<uint> Indices,
     IReadOnlyList<UiDraw> Draws
 );
